requirements.txt
# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
pybind11>=2.10.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
pre-commit>=2.20.0

# Optional dependencies
opencv-python>=4.6.0
matplotlib>=3.5.0
scikit-image>=0.19.0
torch>=1.12.0
torchvision>=0.13.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Testing and benchmarking
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0
